Refactored so that finding a random user is in createChat method instead of chatActivity
